refs #24461 previous regex wa snot handling scientific notation, generating wrong values
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/EuPathWebSvcCommon/trunk@78977 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ WSFPlugin/src/java/org/eupathdb/websvccommon/wsfplugin/blast/NcbiBlastResultFormatter.java
@@ -153,7 +153,7 @@ public class NcbiBlastResultFormatter extends AbstractResultFormatter {
       String summary = summaries.get(sourceId);
       String evalue = getField(summary, findEvalue(summary));
       int[] scoreLocation = findScore(summary);
-      float score = Float.valueOf(getField(summary, scoreLocation));
+      float score = (float)Float.valueOf(getField(summary, scoreLocation));
 
       // insert a link to the alignment section - need to do it before the id link.
       summary = insertUrl(summary, scoreLocation, "#" + sourceId);
